Sort nodes by slots used when scheduing isolated
Because nimbus removes "dead" slots (slots for which their workers have
not yet sent a heartbeat) before schedule is called, we cannot rely on
teh number of free slots on a node.  This will break for clusters whose
nodes have a heterogenious number of slots configured.

Derive the effective number of hosts by taking the minimum of the
config's value and the number of executors in the topology.

If the user requests the topology be scheduled among a number of hosts,
then retry scheduling when the effective number does not match the
scheduled number.