Do not log errors if we have no room to schedule
Messages about there not being enough room to schedule a topology have
been lowered to INFO level, as they are not errors or warnings.

Messages about not being able to assign specific tasks have been moved
to TRACE level, as they are not errors and should not flood the logs at
ERROR level.